Add Register user API
