--- .coverage
@@ -1,1 +1,0 @@
-!coverage.py: This is a private format, don't read it directly!{"lines":{"C:\\Users\\Emmanuel\\SendIt\\API\\__init__.py":[1],"C:\\Users\\Emmanuel\\SendIt\\API\\views.py":[1,2,3,4,5,6,9,11,13,14,15,16,19,23,52,75,165,166,174,175,188,189,198,21,80,81,83,84,85,86,87,88,89,90,93,94,95,96,97,98,100,101,102,105,107,109,111,114,115,117,119,121,123,125,127,129,130,132,133,112,108,110,106,134,138,141,144,148,152,155,158,161,163],"C:\\Users\\Emmanuel\\SendIt\\API\\orders.py":[1,2,5,6,8,9,10,11,12,13,14,15,16,18],"C:\\Users\\Emmanuel\\SendIt\\API\\users.py":[1,4,6,14,17,20,23,26,29,32,35,38,41,44,47,55,64,7,8,9,10,11,45,12],"C:\\Users\\Emmanuel\\SendIt\\API\\data.py":[1,2,3,4,6,7,10,12,19,37,45,63,14,17,65,78,67,68,71],"C:\\Users\\Emmanuel\\SendIt\\API\\responses.py":[1,3,4,7,8,11,12,15,16,19,20,23,24,27,28,30,31]}}
