--- .coveralls.yml
@@ -1,3 +1,0 @@
-service_name: travis-pro
-repo_token: lAirZWlYdy3pA4Nj8RzcoPlc83Z4dVHAQ
-
