+++ .gitignore
@@ -2,3 +2,4 @@ venv/
 __pycache__/
 /__pycache__/
 .pytest_cache/
+ .vscode/
