+++ .travis.yml
@@ -1,6 +1,6 @@
 language: python
 python:
-  - "3.6"
+  - "3.4"
 
 install:
   - pip install -r requirements.txt
