+++ app.py
@@ -1,5 +1,5 @@
-from API.views import app
+from api.views import app
 
 """Runs the app"""
 if __name__ == '__main__':
-    app.run(debug = True) 
+    app.run(debug=True)
