+++ requirements.txt
@@ -22,3 +22,4 @@ pytest==4.0.0
 pytest-cov==2.6.0
 six==1.11.0
 Werkzeug==0.14.1
+psycopg2
